# Conflicts:
#	Assets/StreamingAssets/lang/es-ES.json
#	Assets/StreamingAssets/lang/fr-FR.json
#	Assets/StreamingAssets/lang/it-IT.json
#	Assets/StreamingAssets/lang/ja-JP.json
#	Assets/StreamingAssets/lang/pl-PL.json
#	Assets/StreamingAssets/lang/pt-BR.json
#	Assets/StreamingAssets/lang/pt-PT.json

+++ Assets/Script/Venue/BundleBackgroundManager.cs
@@ -1,14 +1,16 @@
 using System;
+using System.Collections.Generic;
 using System.IO;
 using UnityEngine;
 using YARG.Gameplay;
+using YARG.Helpers;
 using YARG.Venue.VenueCamera;
+using YARG.Venue.Characters;
 
 #if UNITY_EDITOR
 using System.Linq;
 using UnityEditor;
 using UnityEditor.Animations;
-using YARG.Venue.Characters;
 #endif
 
 namespace YARG.Venue
@@ -17,8 +19,11 @@ namespace YARG.Venue
     {
         // DO NOT CHANGE THIS! It will break existing venues
         public const string BACKGROUND_PREFAB_PATH = "Assets/_Background.prefab";
+        public const string CHARACTER_PREFAB_PATH = "Assets/_Character.prefab";
         public const string BACKGROUND_SHADER_BUNDLE_NAME = "_metal_shaders.bytes";
+        public const string CHARACTER_SHADER_BUNDLE_NAME = "_character_metal_shaders.bytes";
         public const string BACKGOUND_OSX_MATERIAL_PREFIX = "_metal_";
+        public const string BUNDLE_OSX_SUFFIX = "_metal.bytes";
 
         private const string VENUE_LAYER_NAME = "Venue";
 
@@ -31,8 +36,13 @@ namespace YARG.Venue
         [SerializeField]
         private Camera mainCamera;
 
-        public AssetBundle Bundle { get; set; }
-        public AssetBundle ShaderBundle { get; set; }
+        [Space]
+        [SerializeField]
+        private VenueCharacter _replaceableVocalist;
+
+        public AssetBundle       Bundle        { get; set; }
+        public List<AssetBundle> ShaderBundles { get; set; } = new();
+        public List<AssetBundle> CharacterBundles { get; set; } = new();
 
         private void Awake()
         {
@@ -69,10 +79,35 @@ namespace YARG.Venue
 
         private void OnDestroy()
         {
-            Bundle.Unload(true);
-            if (ShaderBundle != null)
+            if (Bundle != null)
             {
-                ShaderBundle.Unload(true);
+                Bundle.Unload(true);
+            }
+
+            if (ShaderBundles.Count > 0)
+            {
+                foreach (var bundle in ShaderBundles)
+                {
+                    if (bundle != null)
+                    {
+                        bundle.Unload(true);
+                    }
+                }
+
+                ShaderBundles.Clear();
+            }
+
+            if (CharacterBundles.Count > 0)
+            {
+                foreach (var bundle in CharacterBundles)
+                {
+                    if (bundle != null)
+                    {
+                        bundle.Unload(true);
+                    }
+                }
+
+                CharacterBundles.Clear();
             }
         }
 
@@ -87,11 +122,13 @@ namespace YARG.Venue
 
         private GameObject _backgroundReference;
 
-        [ContextMenu("Export Background")]
-        public void ExportBackground()
+        private void Export(GameObject root, BackgroundHelper.ExportType type)
         {
-            _backgroundReference = gameObject;
-            string path = EditorUtility.SaveFilePanel("Save Background", string.Empty, "bg", "yarground");
+            _backgroundReference = root;
+            string defaultName = type == BackgroundHelper.ExportType.Character ? "character" : "bg";
+            string extension = type == BackgroundHelper.ExportType.Character ? "yargchar" : "yarground";
+            string title = type == BackgroundHelper.ExportType.Character ? "Export Character" : "Export Background";
+            string path = EditorUtility.SaveFilePanel(title, string.Empty, defaultName, extension);
 
             GameObject clonedBackground = null;
 
@@ -103,6 +140,8 @@ namespace YARG.Venue
                 {
                     return;
                 }
+                string fileName = Path.GetFileName(path);
+                string folderPath = Path.GetDirectoryName(path);
 
                 // First we'll collect all shaders and build a separate bundle out of them
                 // for Mac as no other build target will include Metal shaders
@@ -110,6 +149,7 @@ namespace YARG.Venue
 
                 // We use materials as "anchors" to make sure all required
                 // shader variants are included
+                var metalAssetBundleName = fileName + BUNDLE_OSX_SUFFIX;
                 var materialAssets = EditorUtility.CollectDependencies(new[] { gameObject })
                     .OfType<Material>() // Only material dependencices
                     .Select((mat, i) =>
@@ -143,7 +183,7 @@ namespace YARG.Venue
                 if (materialAssets.Length > 0)
                 {
                     var metalAssetBundleBuild = default(AssetBundleBuild);
-                    metalAssetBundleBuild.assetBundleName = BACKGROUND_SHADER_BUNDLE_NAME;
+                    metalAssetBundleBuild.assetBundleName = metalAssetBundleName;
                     metalAssetBundleBuild.assetNames = materialAssets.Concat(shaderAssets).ToArray();
 
                     BuildPipeline.BuildAssetBundles(Application.temporaryCachePath,
@@ -153,10 +193,17 @@ namespace YARG.Venue
                         }, BuildAssetBundleOptions.ForceRebuildAssetBundle,
                         BuildTarget.StandaloneOSX);
 
-                    var filePath = Path.Combine(Application.temporaryCachePath, BACKGROUND_SHADER_BUNDLE_NAME);
-                    var assetPath = Path.Combine(Application.dataPath, BACKGROUND_SHADER_BUNDLE_NAME);
+                    var filePath = Path.Combine(Application.temporaryCachePath, metalAssetBundleName);
+
+                    if (!File.Exists(filePath))
+                    {
+                        EditorUtility.DisplayDialog("Export Unsuccessful", "Failed to build MacOS Shader bundle. See console for more info.", "OK");
+                        throw new FileNotFoundException("MacOS Shader bundle failed to build. <a href=\"https://wiki.yarg.in/wiki/Venue_Creation\">Please ensure you have the \"MacOS Build Support (Mono)\" module installed.</a>");
+                    }
+
+                    var assetPath = Path.Combine(Application.dataPath, metalAssetBundleName);
                     File.Move(filePath, assetPath);
-                    AssetDatabase.ImportAsset(Path.Combine("Assets", BACKGROUND_SHADER_BUNDLE_NAME));
+                    AssetDatabase.ImportAsset(Path.Combine("Assets", metalAssetBundleName));
                 }
                 // Now delete our material clones
                 foreach (var assetPath in materialAssets)
@@ -166,13 +213,12 @@ namespace YARG.Venue
 
                 clonedBackground = Instantiate(_backgroundReference.gameObject);
 
-                string fileName = Path.GetFileName(path);
-                string folderPath = Path.GetDirectoryName(path);
+                var backgroundPath = BackgroundHelper.ExportType.Character == type ? CHARACTER_PREFAB_PATH : BACKGROUND_PREFAB_PATH;
 
                 var assetPaths = new[]
                 {
-                    Path.Combine("Assets/", BACKGROUND_SHADER_BUNDLE_NAME),
-                    BACKGROUND_PREFAB_PATH
+                    Path.Combine("Assets/", metalAssetBundleName),
+                    backgroundPath
                 };
 
                 AssetBundleBuild assetBundleBuild = default;
@@ -214,7 +260,7 @@ namespace YARG.Venue
                     character.LayerStates = layerStates;
                 }
 
-                PrefabUtility.SaveAsPrefabAsset(clonedBackground.gameObject, BACKGROUND_PREFAB_PATH);
+                PrefabUtility.SaveAsPrefabAsset(clonedBackground.gameObject, backgroundPath);
 
                 BuildPipeline.BuildAssetBundles(Application.temporaryCachePath,
                     new[]
@@ -251,6 +297,20 @@ namespace YARG.Venue
                     DestroyImmediate(clonedBackground);
                 }
             }
+        }
+
+        [ContextMenu("Export Vocalist")]
+        public void ExportCharacter()
+        {
+            var vocalist = _replaceableVocalist.gameObject;
+
+            Export(vocalist, BackgroundHelper.ExportType.Character);
+        }
+
+        [ContextMenu("Export Background")]
+        public void ExportBackground()
+        {
+            Export(gameObject, BackgroundHelper.ExportType.Background);
         }
 #endif
     }
